Add option to prepare the boot partition
The boot partition can be prepared with required boot files in order to boot a specific project on a specific board.
The PROJECT_NAME and BOARD can be set in the config file.
Project description is retrieved from kuiper.json.
The configuration is done in 07.export-stage/03.export-image.
After built, Kuiper will be ready to boot with no additional configuration.

Signed-off-by: mlradu <larisa.radu@analog.com>