06.boot-partition: Change /boot to /boot/firmware
To comply with Debian convention, Raspberry Pi OS has transitioned to mounting the boot partition
to /boot/firmware instead of /boot. This commit adjusts all boot partition paths to the new
convention. The bootloader process is unaffected, because it simply searches for the first FAT
partition on disk, regardless of where linux later mounts it.

Review: does this break Kuiper's configure-setup.sh flow for changing targets?

https://github.com/raspberrypi/documentation/issues/3089

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>