03.system-tweaks: Add ischroot=true hack
Add an optional "hack" that replaces ischroot with /usr/bin/true for the duration
of the build, because we can safely assume all commands run during the build are
done in chroot. Select with `HACK_ISCHROOT_TRUE=y`.

Debtools provides `ischroot` as an utility to chech if one is running in a chroot
environment. This is used in various package installations to avoid behaviour
that is invalid inside a chroot. The current version doesn't function properly in
our build environment, returning 2 (= chroot detection is not possible), which
breaks some package installs (e.g. raspi-bootloader, but not only).

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>