01.bootstrap: Don't update-binfmts if running on an arm host
Add an if clause to only activate the `qemu-arm` binfmt on non-arm
hosts. On arm hosts, no emulation is needed. Attempting to activate
the arm binfmt on native arm hosts leads to an error because the
`qemu-user-binfmt` package doesn't contain binfmt configs for their
respective host architecture (i.e. the amd64 package doesn't contain
qemu-x86_64, the arm64 package doesn't contain qemu-arm, etc.).

This makes it possible to build Kuiper images on an arm64 host,
allowing for build speedups by removing the emulation layer.

It should also allow building of 32 bit images on armv6/v7/v7l hosts.

Building armhf images on an arm64 host
--------------------------------------

Note that building armhf images requires the host run a kernel with
4k pages, such as `kernel8.img` instead of `kernel2712.img` on the
Raspberry Pi 5. The kernel included in the image can be anything.
On a 16k page kernel, many binaries will fail to load inside the chroot
because of page alignment. The opposite is not an issue: binaries with
16k pages are compatible with a kernel with 4k pages.

See: https://github.com/raspberrypi/bookworm-feedback/issues/107

Signed-off-by: Ioan Dragomir <ioan.dragomir@analog.com>